tests with local file references
These tests have local file references with a catalogue that translates
systemIds. None of the tests succeed - it might be the case that only
some of the tests are actually conformant with the expected arguments,
but it is certainly the case that Saxon (for example) will call with
such arguments.