Make the setters public on ResourceResponseImpl
Now that the API is defined in terms of ResourceRequest and ResourceResponse
interfaces, it seems unnecessary to keep the setters package-private on the
implementation classes. Anyone reaching down to the implementation classes
presumably knows what they're doing and forcing users to re-implement the
interface just to access the setters seems churlish.

Fix #220