Throw exception of access is denied
Previous versions of the resolver would return null if access to a resource was
denied (if, for example, only http: and https: URIs were allowed but a file: URI
was requested). Unfortunately, returning null may be interpreted by the parser
as a request to simply load the specified resource, bypassing the prohibition.

The only reliable way to avoid this problem is to throw an exception.

This PR also updates a few dependencies and fixes the debugging output issue.

Fix #224