Refactored the “always resolve” feature to avoid namespaces
Fix #229

There’s an unfortunate interaction between the ResolverFeature.ALWAYS_RESOLVE
feature and namespace URI lookup. Using a namespace URI to find, for example, a
schema in the catalog, or via indirection with RDDL to find things online, makes
sense. But, generally speaking, what’s located at the namespace URI is not a
useful resource. It’s very often not even present.

This PR refactors how the “always resolve” feature is implemented so that it can
be set on a per-request basis. By default, it’s the same as the global
configuration setting, but when doing namespace URI lookup, we disable it.