Assure that responses are resolved against the request base URI
Fix #248

If we find ourselves trying to get a resource from a response and the URI isn’t
absolute, and the base URI of the request also isn’t absolute, resolve the base
URI of the request against the current working directory, then the relative URI
against that. This is more likely to be correct.