Use unmasked URIs for RDDL; also, disable RDDL parsing by default
Fix #253

By default jar: URIs are masked. If a resource is resolved locally, but the
resolved location is in a jar: file, we return the original URI not the resolved
URI. This is necessary because Java.net.URI doesn’t understand jar: URIs. (Oh,
the irony.)

Unfortunately, when the RDDL parser asked for the resolved URI, it got back the
original too. So it would go attempt to parse the “real” location not the
resolved resource. That’s been fixed. It now parses the unmasked URI, not the
resolved URI.

This is also the second bug (I think) caused by RDDL parsing. With some regret,
I’m going to disable RDDL parsing by default. That means it will get used far,
far less often, but since it’s not widely supported, that probably won’t have a
huge impact. If you’re relying on it, remember to enable it.